requirements.txt
asgiref==3.8.1
Django==5.0.6
django-allauth==0.63.1
django-cors-headers==4.3.1
django-rest-auth==0.9.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
